Update AdobeCreativeCloudInstaller.pkg.recipe
Change to copy the DMG itself a resource and have the postinstall script mount it and run the installer within. This is because on Silicon, Adobe's installer app fails if it's copied directly as a package resource (it succeeds when run from within the source DMG).

Because we unmount the DMG at the end, we determine the exit code of Adobe's installer and use this as the postinstall exit code to ensure the package fails too if necessary (e.g. when installing on the wrong CPU architecture).